requirements.txt
# python-prctrl>=1.7
colorama>=0.3.9
# notify2>=0.3.1
# dbus-python>=1.2.8
websocket-client>=0.44.0
requests>=2.20.0
# pandas>=0.23.4
# numpy>=1.15.1
# scipy>=1.1.0
# statistics>=1.0.3.5
dateparser>=0.7.0
pytz>=1.5.1
autobahn>=18.9.2
twisted>=18.7.0
pyOpenSSL>=18.0.0
service_identity>=17.0.0
psycopg2>=2.7.3.2
# mpl_finance>=0.10.0
# tabulate>=0.8.3
# TA-Lib>=0.4.17
# for the rpc server of connectors
# asyncrpc>=0.1.4
pyzmq>=1.5
hiredis>=1.0.0
